moving date range picker to individual charts, adding reset btn to filter
