+++ src/dashboard/src/components/charts/bar/allocationByStorageVolume/AllocationByStorageVolume.module.scss
@@ -122,7 +122,7 @@
   }
 }
 
-.footer {
+.allocationFooter {
   display: flex;
   position: absolute;
   bottom: 28px;
