+++ src/dashboard/src/components/charts/bar/smallBar/SmallBarChart.module.scss
@@ -116,6 +116,7 @@
     width: 100%;
     border-radius: 10px;
     margin-top: 5px;
+    overflow: hidden;
 
     .percentage {
         height: 20px;
