+++ src/dashboard/src/components/charts/line/LineChart.module.scss
@@ -17,3 +17,16 @@
         min-height: 420px;
     }
 }
+
+.date {
+    margin-top: 10px;
+    display: flex;
+    gap: 1rem;
+
+    button {
+        position: relative;
+        padding: 5px 10px;
+        top: unset;
+        right: unset;
+    }
+}
