+++ src/dashboard/src/components/header/Header.module.scss
@@ -34,7 +34,8 @@
   }
 
   &.filterPadding {
-    margin-bottom: 88px;
+    margin-bottom: 78px;
+    min-height: 260px;
   }
 
   &.adminPadding {
